Fix: integration tests were not executed
