+++ .travis.yml
@@ -1,7 +1,7 @@
 env:
   global:
    # The next declaration is the encrypted COVERITY_SCAN_TOKEN, created
-   #   via the "travis encrypt" command using the project repo's public key
+   # via the "travis encrypt" command using the project repo's public key
    - secure: "rumJzPX8ade8PN3svS4HcVNNFRh0/wbAFHaaZuFARiS0FKDnWQ5xUl7I+xeE6+PI4r5hZJWH6C74wgvbTwS2OsSrruBYR4l7YjGneu8YRQwYuCxsS66X1Pslss1ykqEuuX6+pk5dovhve8oAO3jR3Jkm4pdC0kSAVsnpYjbC7mU4s916h5i2reYt2VamZ+2HCfNlOp8yKLZSNWxubp7FaV0O1qx9g9SIHnvO/m/aoH+NHu0lZh9Sdw7YrjV22ESyrd7cyI93TshVHiswNDwIZcahfgH0I81SNHeuIeARiBVr/2495obMbR2WC/3BrdfeMekpoipFIoWf18s/Bp5KmfZJaxeObRiT+LL3dgqIKVyz2bT7wQIwIVNShODO0zjoLK7hO+7WSEMLMvpoUEuprO9037Mbn20TkCbsTlax0qQeh0Oe+sgLLjxyuTTIf5m+ObwgsEAe05+yaY60A51w20iva8A+qAKbbf08gBBZEf7PKY4XZF4/T44BaXWabbVPBI8XOT8UWLFxc1jIb5T6feVIcR6y30b8N0145BJbW1AXJjpg/BS05+AmqYJXkNUwSGMj/SaS5KMXsSPCsnfVtCZaYmUlT8QbGYVXuauCpDvQpESWHEkPmm6tXCt41d+jYMDMxaPokzRI90s/CbfhDYRiUMeXaT11/3v6ZTo4458="
 
 language: java
@@ -11,7 +11,8 @@ jdk:
   - oraclejdk8
   - openjdk7
 
-script: mvn clean checkstyle:check package
+# use 'verify' phrase to execute integration tests
+script: mvn clean checkstyle:check verify
 
 after_success:
   bash <(curl -s https://codecov.io/bash)
