+++ src/test/java/pl/szczepanik/silencio/integration/GeoLocationConverterIntegrationTest.java
@@ -10,7 +10,7 @@ import org.junit.Test;
 import pl.szczepanik.silencio.GenericTest;
 import pl.szczepanik.silencio.api.Format;
 import pl.szczepanik.silencio.api.Processor;
-import pl.szczepanik.silencio.converters.WikipediaConverter;
+import pl.szczepanik.silencio.converters.GeoLocationConverter;
 import pl.szczepanik.silencio.core.Builder;
 import pl.szczepanik.silencio.utils.JSONUtility;
 import pl.szczepanik.silencio.utils.ResourceLoader;
@@ -18,10 +18,10 @@ import pl.szczepanik.silencio.utils.ResourceLoader;
 /**
  * @author Damian Szczepanik (damianszczepanik@github)
  */
-public class WikipediaConverterTestInt extends GenericTest {
+public class GeoLocationConverterIntegrationTest extends GenericTest {
 
     @Test
-    public void shouldGetWordsFromWikipediaOrg() throws IOException {
+    public void shouldGetWordsFromGoogleApi() throws IOException {
         // WARNING: this test fail when no Internet connection is available
 
         // given
@@ -29,7 +29,7 @@ public class WikipediaConverterTestInt extends GenericTest {
         output = new StringWriter();
 
         // when
-        Processor processor = new Builder(Format.JSON).with(new WikipediaConverter()).build();
+        Processor processor = new Builder(Format.JSON).with(new GeoLocationConverter()).build();
         processor.load(input);
         processor.process();
         processor.write(output);
